+++ symptoms/src/main/java/com/protean/symptoms/repository/PriscriptionRepository.java
@@ -1,5 +1,6 @@
 package com.protean.symptoms.repository;
 
+import java.math.BigInteger;
 import com.protean.symptoms.entity.PriscriptionDtlsEntity;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -11,7 +12,7 @@ public interface PriscriptionRepository extends CrudRepository<PriscriptionDtlsE
 
 	
 	@Query(value = "SELECT * FROM public.priscription_dtls where uniqueid= :uniqueId", nativeQuery = true)
-	PriscriptionDtlsEntity findByuniqueId(@Param("uniqueId") Integer uniqueId);
+	PriscriptionDtlsEntity findByuniqueId(@Param("uniqueId") BigInteger uniqueId);
 
 	List<PriscriptionDtlsEntity> findAll();
